requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.1.1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-dbbackup==3.3.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
stripe==2.55.0
urllib3==1.25.11
webencodings==0.5.1